requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.4
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
python-dotenv==0.13.0
six==1.15.0
typing-extensions==3.7.4.2
websockets==8.1
yarl==1.4.2
youtube-dl==2020.6.16.1
